requirements.txt
accelerate==0.18.0
diffusers==0.21.0
hydra-core==1.3.2
imageio==2.31.5
scikit-image==0.22.0
timm==0.9.7
torch-ema==0.3
transformers==4.26.0
ninja
opencv-python-headless
plotly
rich
scipy
tqdm
wandb
trimesh
open3d
ipdb